setup initial test infra for pureport-client (#43)
This commit enables a test infra for testing the pureport-client
package.  The test infra is executed using pytest and orchestrated
locally using tox.  For pull requets and pushs, the test orchestrator is
Github Actions.

It is important to note that for the moment the two test orchestrators
are independent and therefore any changes to one *must* be propogated to
the other one.

To run tests from a local development environment, simply using the make
target ```make test```

Signed-off-by: Peter Sprygada <peter.sprygada@pureport.com>