refactor the exceptions module
This commit refactors the exceptions modules to simplify the interface
and number of support exception classes.  It introduces an exception
hierarchy with all exceptions now subclasses of `PureportClientError`.

The changes in this commit also remove generating status specific
exceptions based on http status code in lieu of a more streamlined
single exception `ClientHttpError` that contains the status code and
reason.

Signed-off-by: Peter Sprygada <peter.sprygada@pureport.com>