separate client and comamnd tree
This commit separates the Pureport Client instance from the instance
used to run the CLI.  It breaks up the Client object into discrete parts
and dynamically loads the commands to generate the CLI.  This change
will make it substaintially easier to add new commands as well as write
test cases for the existing CLI commands.

This is a signinficant structurally change to how the CLI is generated.
Every attempt has been made to avoid making any changes to the CLI
itself with regards to features.

Signed-off-by: Peter Sprygada <peter.sprygada@pureport.com>