updates client test cases to match command refactor
This commit mirrors the changes introduced in 317c56 with regards to
providing a set of unit test cases for testing the cli commands.  As
before, every effort was made not to change the test methodology, rather
simply streamline the code and provide test coverage.

One significant exception to the above note is that the new unit tests
load the api spec from a file now instead of retrieving it from the
production instance.  This allows the test cases to be fully mocked and
run offline but introduces the opportunity for tests to pass but the
command fail the spec and production api are out on sync.

Signed-off-by: Peter Sprygada <peter.sprygada@pureport.com>